v1/labs/Lab1_Hello_Autogen/requirements.txt
# Lab 1 Dependencies
pyautogen>=0.2.0
python-dotenv>=1.0.0
